Update remollH1.C
In remollAna.C, the fName of the remollH1 object is always set with a plot name plus generator name as suffix (remollAna.C-line 11 for example: Form("rad_%s",gen)). The units for the rate-axis are also defined here based on the binsize which is defined in remollH1.C. The rad_%s for example has units of GHz/5mm in a range of 0.6 to 1.3 m. This unit is accurate if there are 140 bins in that range.
But in remollH1.C(line-13-25), there is no fName that is rad_moller. So, the bin number is always set as 300. So, the height of the distributions are not represented in the correct units. Hence the corrections were made from line 16-23.